feat(home): set title
