+++ web/components/Landing/Title.module.css
@@ -1,0 +1,9 @@
+@font-face {
+  font-family: 'museo';
+  src: url('./museo-500.otf');
+  font-weight: 500;
+}
+
+.title {
+  font-family: 'museo';
+}
